backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
solace-pubsubplus==1.5.0
Werkzeug==3.0.1
